feat: add customization point for decimal
